requirements.txt
turbopuffer>=1.17.0
datasets>=2.14.0,<3.0.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
huggingface_hub>=0.19.0
safetensors>=0.4.0
python-dotenv>=1.0.0
